refactor createMessageThread
