+++ extensions/elation/actions/createMessageThread/config/dataPoints.ts
@@ -1,0 +1,8 @@
+import { type DataPointDefinition } from '@awell-health/extensions-core'
+
+export const dataPoints = {
+  messageThreadId: {
+    key: 'messageThreadId',
+    valueType: 'number',
+  },
+} satisfies Record<string, DataPointDefinition>
