+++ extensions/elation/actions/createMessageThread/config/index.ts
@@ -1,0 +1,2 @@
+export { fields, FieldsValidationSchema } from './fields'
+export { dataPoints } from './dataPoints'
