+++ extensions/elation/actions/createMessageThread/index.ts
@@ -1,0 +1,1 @@
+export { addAllergy } from './createMessageThread'
